requirements.txt
ultralytics
gradio
Pillow
